sys/vfs/hammer: Cast HAMMER_BLOCKMAP_LAYER1|2_INDEX macros to int
This commit does the same as 37fa1bc7 for INDEX macros.

RADIX macro is how many layer entries a big-block can hold,
INDEX macro is an index of layer entry for the given offset.
(i.e. Maximum INDEX equals (RADIX - 1))

+++ sbin/hammer/misc.c
@@ -344,8 +344,8 @@ hammer_add_zone_stat(struct zone_stat *stats, hammer_off_t offset,
 	vol = HAMMER_VOL_DECODE(offset);
 
 	offset &= HAMMER_OFF_SHORT_MASK;  /* cut off volume bits from layer1 */
-	i = (int)HAMMER_BLOCKMAP_LAYER1_INDEX(offset);
-	j = (int)HAMMER_BLOCKMAP_LAYER2_INDEX(offset);
+	i = HAMMER_BLOCKMAP_LAYER1_INDEX(offset);
+	j = HAMMER_BLOCKMAP_LAYER2_INDEX(offset);
 
 	if (i > l1_max[vol]) {
 		assert(i < 1024);  /* XXX hardcoded */
